refactor update os, move os and hath to separate folders
